refactor(oauthProxy): split into lib/oauthProxy/ behind stable facade
Phase 3 of issue #24 follow-up. Splits the 1003-line lib/oauthProxy.ts
into focused modules:
- types.ts, prompts.ts, references.ts, errors.ts
- runtime.ts, streams.ts, generators.ts, index.ts (barrel)

lib/oauthProxy.ts is now a 1-line facade re-exporting from index.ts;
all external imports continue to work unchanged.

Contract tests updated to read concatenated split sources.

No runtime behavior change. 610/610 tests pass.

Refs: #24
