+++ tests/multimode-backend-contract.test.js
@@ -5,7 +5,24 @@ import { join } from "node:path";
 
 const root = process.cwd();
 
+// NOTE: lib/oauthProxy.ts was split into lib/oauthProxy/*.ts behind a facade;
+// readSource("lib/oauthProxy.ts") now returns all split sources concatenated.
+const OAUTH_PROXY_SOURCES = [
+  "lib/oauthProxy.ts",
+  "lib/oauthProxy/types.ts",
+  "lib/oauthProxy/prompts.ts",
+  "lib/oauthProxy/references.ts",
+  "lib/oauthProxy/errors.ts",
+  "lib/oauthProxy/runtime.ts",
+  "lib/oauthProxy/streams.ts",
+  "lib/oauthProxy/generators.ts",
+  "lib/oauthProxy/index.ts",
+];
+
 function readSource(path) {
+  if (path === "lib/oauthProxy.ts") {
+    return OAUTH_PROXY_SOURCES.map((p) => readFileSync(join(root, p), "utf8")).join("\n");
+  }
   return readFileSync(join(root, path), "utf8");
 }
