+++ tests/node-child-refs-contract.test.js
@@ -5,7 +5,19 @@ import { readFileSync } from "node:fs";
 const store = readFileSync("ui/src/store/useAppStore.ts", "utf-8");
 const imageNode = readFileSync("ui/src/components/ImageNode.tsx", "utf-8");
 const nodes = readFileSync("routes/nodes.ts", "utf-8");
-const oauth = readFileSync("lib/oauthProxy.ts", "utf-8");
+// NOTE: lib/oauthProxy.ts was split into lib/oauthProxy/*.ts behind a facade.
+const OAUTH_PROXY_SOURCES = [
+  "lib/oauthProxy.ts",
+  "lib/oauthProxy/types.ts",
+  "lib/oauthProxy/prompts.ts",
+  "lib/oauthProxy/references.ts",
+  "lib/oauthProxy/errors.ts",
+  "lib/oauthProxy/runtime.ts",
+  "lib/oauthProxy/streams.ts",
+  "lib/oauthProxy/generators.ts",
+  "lib/oauthProxy/index.ts",
+];
+const oauth = OAUTH_PROXY_SOURCES.map((p) => readFileSync(p, "utf-8")).join("\n");
 
 describe("child node references contract", () => {
   it("does not block child node references in the frontend", () => {
