+++ tests/node-context-policy.test.js
@@ -3,7 +3,19 @@ import assert from "node:assert/strict";
 import { readFileSync } from "node:fs";
 
 const route = readFileSync("routes/nodes.ts", "utf-8");
-const oauth = readFileSync("lib/oauthProxy.ts", "utf-8");
+// NOTE: lib/oauthProxy.ts was split into lib/oauthProxy/*.ts behind a facade.
+const OAUTH_PROXY_SOURCES = [
+  "lib/oauthProxy.ts",
+  "lib/oauthProxy/types.ts",
+  "lib/oauthProxy/prompts.ts",
+  "lib/oauthProxy/references.ts",
+  "lib/oauthProxy/errors.ts",
+  "lib/oauthProxy/runtime.ts",
+  "lib/oauthProxy/streams.ts",
+  "lib/oauthProxy/generators.ts",
+  "lib/oauthProxy/index.ts",
+];
+const oauth = OAUTH_PROXY_SOURCES.map((p) => readFileSync(p, "utf-8")).join("\n");
 const providerOptions = readFileSync("lib/providerOptions.ts", "utf-8");
 
 describe("node context and edit search policy", () => {
