+++ tests/oauth-masked-edit-contract.test.js
@@ -2,7 +2,19 @@ import fs from "node:fs";
 import assert from "node:assert/strict";
 import { test } from "node:test";
 
-const proxy = fs.readFileSync("lib/oauthProxy.ts", "utf8");
+// NOTE: lib/oauthProxy.ts was split into lib/oauthProxy/*.ts behind a facade.
+const OAUTH_PROXY_SOURCES = [
+  "lib/oauthProxy.ts",
+  "lib/oauthProxy/types.ts",
+  "lib/oauthProxy/prompts.ts",
+  "lib/oauthProxy/references.ts",
+  "lib/oauthProxy/errors.ts",
+  "lib/oauthProxy/runtime.ts",
+  "lib/oauthProxy/streams.ts",
+  "lib/oauthProxy/generators.ts",
+  "lib/oauthProxy/index.ts",
+];
+const proxy = OAUTH_PROXY_SOURCES.map((p) => fs.readFileSync(p, "utf8")).join("\n");
 const cfg = fs.readFileSync("config.ts", "utf8");
 
 test("config exposes oauth.maskedEditEnabled flag", () => {
